refactor: change files structure and naming accordingly to agreed
