+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { FC, useEffect } from 'react';
 import './App.css';
-import Header from './components/Header/header';
-import Content from './components/Content/content';
-import socket from './Api';
+import Header from './components/Header/Header';
+import Content from './components/Content/Content';
+import socket from './api';
 
 const App: FC = () => {
   useEffect(() => {
